Refactor main package into separate files
